Model Deploy/requirements.txt
# pip install -r requirements.txt

tensorflow-gpu==2.1.0
numpy== 1.18.1
pandas==1.0.3
matplotlib== 3.1.3